requirements.txt
qiskit==1.0.1
qiskit-aer==0.13.3
opencv-python==4.9.0.80
numpy==1.26.4
pyserial==3.5
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
supabase==2.28.0
pillow
